Create PYTHON(READ1).11.1023.16:43
translate this program to python....#include <iostream>
#include <fstream>
#include <string>
using namespace std;

int main () {
  string line;
  ifstream myfile ("example.txt");
  if (myfile.is_open())
  {
    while ( getline (myfile,line) )
    {
      cout << line << '\n'; //do work on each line of the file here
    }
    myfile.close();
  }

  else cout << "Unable to open file"; 

  return 0;
}
Sure, here is the translation of the C++ program you provided to Python:

Python
def read_file_line_by_line(filename):
  """Reads a file line by line and returns a list of lines.

  Args:
    filename: The name of the file to read.

  Returns:
    A list of strings, each line of the file.
  """

  lines = []
  with open(filename, "r") as f:
    for line in f:
      lines.append(line)
  return lines


def main():
  filename = "example.txt"
  lines = read_file_line_by_line(filename)

  for line in lines:
    # do work on each line of the file here
    print(line)


if __name__ == "__main__":
  main()
Use code with caution. Learn more
This Python program works in a similar way to the C++ program. It first opens the file example.txt in read mode. Then, it iterates over the lines of the file and prints each line to the console.

Here is an example of how to use the Python program:

Python
>>> read_file_line_by_line("example.txt")
['This is line 1.', 'This is line 2.', 'This is line 3.']
>>>
Use code with caution. Learn more
profile picture
REWRITE THE ABOVE PROGRAM BUT ALLOW USER TO INPUT THE NAME OF THE FILE TO READ INTO
View other drafts









Sure, here is a rewritten version of the Python program that allows the user to input the name of the file to read:

Python
import sys


def read_file_line_by_line(filename):
  """Reads a file line by line and returns a list of lines.

  Args:
    filename: The name of the file to read.

  Returns:
    A list of strings, each line of the file.
  """

  lines = []
  with open(filename, "r") as f:
    for line in f:
      lines.append(line)
  return lines


def main():
  filename = input("Enter the name of the file to read: ")

  lines = read_file_line_by_line(filename)

  for line in lines:
    # do work on each line of the file here
    print(line)


if __name__ == "__main__":
  main()
Use code with caution. Learn more
This program works in the same way as the previous Python program, except that it prompts the user for the name of the file to read at the beginning of the program.

Here is an example of how to use the rewritten Python program:

Python
Enter the name of the file to read: example.txt

This is line 1.
This is line 2.
This is line 3*******************************************************************************************************************************************************
#include <iostream>
#include <fstream>
#include <string>
using namespace std;

int main () {
  string line;
  ifstream myfile ("example.txt");
  if (myfile.is_open())
  {
    while ( getline (myfile,line) )
    {
      cout << line << '\n'; //do work on each line of the file here
    }
    myfile.close();
  }

  else cout << "Unable to open file"; 

  return 0;
}