Create SIZEIS0BYTES.11.10.23.17:35
LINE
string line;
if (myfile.is_open())
{
 while (getline (myfile,line)  )
  cout << line
******************************
1cin is an object of class istream and cout is an object of class ostream.
2his code creates a file called example.txt and inserts a sentence into 
it in the same way we are used to do with cout, but using the 
file stream myfile instead.
3ios::in	Open for input operations.
ios::out	Open for output operations.
ios::binary	Open in binary mode.
4myfile.open ("example.bin", ios::out | ios::app | ios::binary);
5Non-binary files are known as text files,
6int main () {
  ofstream myfile ("example.txt");
  if (myfile.is_open())
  {
    myfile << "This is a line.\n";
    myfile << "This is another line.\n";
    myfile.close();
  }
  else cout << "Unable to open file";
  return 0;
}
[file example.txt]
This is a line.
This is another line.
 Edit & run on cpp.sh
7Checking state flags
The following member functions exist to check for specific states of a stream (all of them return a bool value):

bad()
Returns true if a reading or writing operation fails. For example, in the case that we try to write to a file that is not open for writing or if the device where we try to write has no space left.
fail()
Returns true in the same cases as bad(), but also in the case that a format error happens, like when an alphabetical character is extracted when we are trying to read an integer number.
eof()
Returns true if a file open for reading has reached the end.
good()
It is the most generic state flag: it returns false in the same cases in which calling any of the previous functions would return true. 
8he member function clear() can be used to reset the state flags.

get and put stream positioning
All i/o streams objects keep internally -at least- one internal position:

ifstream, like istream, keeps an internal get position with the location of the element to be read in the next input operation.

ofstream, like ostream, keeps an internal put position with the location where the next element has to be written.

Finally, fstream, keeps both, the get and the put position, like iostream.
9tellg() and tellp()
These two member functions with no parameters return a value of the member type streampos, which is a type representing the current get position (in the case of tellg) or the put position (in the case of tellp).

seekg() and seekp()
These functions allow to change the location of the get and put positions. Both functions are overloaded with two different prototypes. The first form is:

seekg ( position );
seekp ( position );

Using this prototype, the stream pointer is changed to the absolute position position (counting from the beginning of the file). The type for this parameter is streampos, which is the same type as returned by functions tellg and tellp.

The other form for these functions is:

seekg ( offset, direction );
seekp ( offset, direction );
10// obtaining file size
#include <iostream>
#include <fstream>
using namespace std;

int main () {
  streampos begin,end;
  ifstream myfile ("example.bin", ios::binary);
  begin = myfile.tellg();
  myfile.seekg (0, ios::end);
  end = myfile.tellg();
  myfile.close();
  cout << "size is: " << (end-begin) << " bytes.\n";
  return 0;
}
size is: 40 bytes.
 Edit & run on cpp.sh

**************************************************************
s "isopen()" in the c++ standard library as a built in function in...if (myfile.is_open())
Yes, the is_open() function is in the C++ standard library as a built-in function. It is a member of the ifstream class, which is used to read data from files. The is_open() function returns true if the file is open, and false if the file is not open.

The if (myfile.is_open()) statement in the C++ program you provided checks if the file example.txt is open. If the file is open, the program will continue to execute the code inside the loop. Otherwise, the program will print the message "Unable to open file" to the console.

Here is a description of the different parts of the if (myfile.is_open()) statement:

if: This keyword marks the beginning of the conditional statement.
myfile.is_open(): This function call checks if the file example.txt is open.
():: These parentheses enclose the function call.
{}: These curly braces enclose the code that will be executed if the file is open.
The is_open() function is a useful way to check if a file is open before trying to read or write data to the file. This can help to prevent errors from occurring.

profile picture
GIVE A SIMPLE EXAMPLE OF A DIFFERENTIAL EQUATION WITH THE SOLUTION
View other drafts





























A simple example of a differential equation is:

dy/dx = y
This equation states that the rate of change of the variable y is equal to itself. The solution to this equation is:

y = Ce^x
where C is an arbitrary constant. This can be verified by substituting the solution back into the original equation:

dy/dx = C e^x = y
Therefore, the solution is correct.

Another example of a differential equation is:

dy/dx = 2x
This equation states that the rate of change of the variable y is equal to twice the value of the variable x. The solution to this equation is:

y = x^2 + C
where C is an arbitrary constant. This can also be verified by substituting the solution back into the original equation:

dy/dx = 2x = 2x + C
Therefore, the solution is also correct.

These are just two simple examples of differential equations with their solutions. There are many other more complex differential equations that are used in mathematics, physics, and engineering.