Create Getline5
Getline from File to Array
For this lab, we will pull a line from a file and store it in an array.  To begin, right click the link to save the example.txt Download example.txtfile in your project directory.

To get a line from a file, use the getline() function, which takes two arguments: the name of the filestream object and a variable to hold the line once it is read.  In the example below, can you find these instructions?

Declare a filestream object named myfile  
Declare a string variable named line
Call the getline function and pass myfile and line
Notice that the call to getline is in the parentheses of the while loop?  That is because getline will return true if there is a line to get and it will return false otherwise.  So, you can call the getline function wherever a true or false is expected.  In this case, we call it in the condition of a while loop.  As long as there is a line to get, the loop will continue.  As soon as it runs out of lines, the loop automatically ends.

Next, we need to add the line variable to the array.  In the example below, can you find these instructions?

Declare an array named myarray
Declare a variable to keep track of the next available index number in the array, name it numused and initialize to zero. 
Load the line variable into myarray at position numused and increment numused
Did you notice that the numused variable is incremented with ++ inside the brackets of the array?  This is just a short cut and is the equivalent of these two lines of code:

myarray[numused] = line; 
numused++
You can do it either way. Whether you decide to use two statements or combine them into one, both operations need to happen.

In this example, a single variable is used to hold all lines of the file. This works because each time through the loop, the line variable is overwritten by the next line of the file. Can you find the line of code that does that?

Example
#include <fstream>
#include <string>
using namespace std;

void myLines();
int main()
{
   myLines();
}

void myLines()
{
  string line;
  string myarray[10];
  int numused=0;
  ifstream myfile ("example.txt");
  if (myfile.is_open())
  {
    while (getline(myfile, line))
     {
      myarray[numused++] = line; 
     }
    myfile.close();
  }
}