Create The line myarray[numused++] = line;
HOW DOES  myarray[numused++] = line; READ LINE FROM FILE IN...#include <fstream>
#include <string>
using namespace std;

void myLines();
int main()
{
   myLines();
}

void myLines()
{
  string line;
  string myarray[10];
  int numused=0;
  ifstream myfile ("example.txt");
  if (myfile.is_open())
  {
    while (getline(myfile, line))
     {
      myarray[numused++] = line; 
     }
    myfile.close();
  }
}
View other drafts


















The line myarray[numused++] = line; in the code you provided reads a line from the file example.txt and stores it in the array myarray. The numused++ operator increments the variable numused by 1, which means that the next line read from the file will be stored in the next element of the array.

The getline() function reads a line from the file myfile and stores it in the string variable line. The while loop continues to read lines from the file until the end of the file is reached.

Here is a breakdown of the code:

C++
#include <fstream>
#include <string>
using namespace std;

void myLines();
int main()
{
   myLines();
}

void myLines()
{
  string line;
  string myarray[10];
  int numused=0;
  ifstream myfile ("example.txt");
  if (myfile.is_open())
  {
    while (getline(myfile, line))
     {
      myarray[numused++] = line; 
     }
    myfile.close();
  }
}
Use code with caution. Learn more
#include <fstream> and #include <string> are header files that contain the definitions of the ifstream and string classes, respectively.
using namespace std; makes all of the names in the std namespace available to the code without having to prefix them with std::.
void myLines(); declares a function called myLines().
int main() is the entry point for the program.
myLines(); calls the myLines() function.
void myLines() is the implementation of the myLines() function.
string line; declares a string variable called line.
string myarray[10]; declares an array of 10 strings called myarray.
int numused=0; declares an integer variable called numused and initializes it to 0.
ifstream myfile ("example.txt"); opens the file example.txt for reading.
if (myfile.is_open()) checks if the file was opened successfully.
while (getline(myfile, line)) reads a line from the file and stores it in the line variable. The while loop continues to read lines from the file until the end of the file is reached.
myarray[numused++] = line; stores the line read from the file in the next element of the myarray array. The numused++ operator increments the variable numused by 1, which means that the next line read from the file will be stored in the next element of the array.
myfile.close(); closes the file example.txt.
Once the myLines() function has finished running, the myarray array will contain all of the lines from the file example.txt.