Create reversepoem.11.12.23.15:25
Here are the features you identified in the C++ code snippet:

Function with a return value:

C++
string inputfilename() {
  string s;

  cout << "Enter filename: ";
  getline(cin, s);
  return s;
}
Use code with caution. Learn more
This function returns a string, which is the filename entered by the user.

Function with a reference parameter:

C++
void reverseme(string v[], int &numused)
Use code with caution. Learn more
This function takes a reference to an integer variable, numused, as a parameter. This allows the function to modify the value of numused in the calling function.

Function with an array as a parameter:

C++
void tovector(string inputfilename, string v[], int &numused)
Use code with caution. Learn more
This function takes an array of strings, v, and an integer variable, numused, as parameters. The function reads lines from the file specified by inputfilename and stores them in the array v. The variable numused keeps track of the number of elements in the array.

For loop embedded in an if clause:

C++
for (int i = numused-1; i>=0; --i) {
  // ...
}
Use code with caution. Learn more
This for loop is embedded in an if clause to check if the file is open. If the file is open, the loop iterates backwards through the array v and writes the lines to the file.

If clause embedded in a loop:

C++
while(getline(in, line)) {
  if (in.fail()) {
    // ...
  } else {
    // ...
  }
}
Use code with caution. Learn more
This if clause is embedded in a while loop to read lines from the file specified by inputfilename. If a line is read successfully, the line is stored in the array v. If an error occurs, the program prints an error message.

Input file stream:

C++
ifstream in(inputfilename);
Use code with caution. Learn more
This statement creates an input file stream object, in, which is used to read lines from the file specified by inputfilename.

Output file stream:

C++
ofstream myfile;
myfile.open("reverse.txt");
Use code with caution. Learn more
This statement creates an output file stream object, myfile, which is used to write lines to the file reverse.txt.

Getline as a condition of a loop:

C++
while(getline(in, line)) {
  // ...
}
Use code with caution. Learn more
The getline() function is used to read a line from the input file stream in. The while loop iterates while the getline() function is successful.

Reading lines of a file to write to an array:

C++
while(getline(in, line)) {
  v[numused++] = line;
}
Use code with caution. Learn more
This while loop reads lines from the input file stream in and stores them in the array v. The variable numused keeps track of the number of elements in the array.

Reading lines from an array to write to a file in the program:

C++
for (int i = numused-1; i>=0; --i) {
  myfile << v[i]<< endl;
}
Use code with caution. Learn more
This for loop iterates backwards through the array v and writes the lines to the output file stream myfile.